calibrate accel with offsets

+++ src/flight/sensors/imu/src/mpu6050_init.h
@@ -19,7 +19,7 @@ bool mpu6050_set_fs_sel_gyro(mpu6050_fs_sel_gyro_t fs_select);
 
 bool mpu6050_calc_gyro_offsets(int16_t *xg_offset, int16_t *yg_offset, int16_t *zg_offset);
 
-bool mpu6050_calc_accel_offsets(float *xa_offset, float *ya_offset, float *za_offset);
+bool mpu6050_calc_accel_offsets(int16_t *xa_offset, int16_t *ya_offset, int16_t *za_offset);
 
 bool mpu6050_set_dlpf(uint8_t config);
